Add unit tests for building the package supplier
