+++ src/main/java/org/spdx/sbom/gradle/utils/SpdxDocumentBuilder.java
@@ -35,9 +35,7 @@ import java.util.Map.Entry;
 import java.util.Optional;
 import java.util.Set;
 import java.util.function.Function;
-import java.util.stream.Collector;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 import javax.annotation.Nullable;
 import org.gradle.api.artifacts.ModuleVersionIdentifier;
 import org.gradle.api.artifacts.component.ComponentArtifactIdentifier;
@@ -64,6 +62,7 @@ import org.spdx.library.model.enumerations.ReferenceCategory;
 import org.spdx.library.model.enumerations.RelationshipType;
 import org.spdx.library.model.license.SpdxNoAssertionLicense;
 import org.spdx.sbom.gradle.extensions.SpdxSbomTaskExtension;
+import org.spdx.sbom.gradle.maven.MavenPackageSupplierBuilder;
 import org.spdx.sbom.gradle.maven.PomInfo;
 import org.spdx.sbom.gradle.project.DocumentInfo;
 import org.spdx.sbom.gradle.project.ProjectInfo;
@@ -352,7 +351,7 @@ public class SpdxDocumentBuilder {
 
       spdxPkgBuilder.setVersionInfo(moduleId.getVersion());
 
-      spdxPkgBuilder.setSupplier(buildMavenPackageSupplier(pomInfo));
+      spdxPkgBuilder.setSupplier(MavenPackageSupplierBuilder.buildPackageSupplier(pomInfo));
 
       String sha1 =
           com.google.common.io.Files.asByteSource(dependencyFile).hash(Hashing.sha1()).toString();
@@ -365,67 +364,6 @@ public class SpdxDocumentBuilder {
 
       return Optional.of(spdxPkgBuilder.build());
     }
-    return Optional.empty();
-  }
-
-  private String buildMavenPackageSupplier(PomInfo pomInfo) {
-    var organizationName =
-        pomInfo
-            .getOrganization()
-            .flatMap(o -> Optional.ofNullable(o.getName()))
-            .map(n -> "Organization: " + n);
-
-    // if all the developers have the same organization, use it as the supplier
-    var developersOrganization =
-        pomInfo.getDevelopers().stream()
-            .map(d -> d.getOrganization().orElse(null))
-            .collect(uniqueOrEmpty())
-            .map(o -> "Organization: " + o);
-
-    // otherwise find the first developer that has a name, or only has an organization
-    // and construct the supplier based on those properties
-    var supplierFromDeveloper =
-        pomInfo.getDevelopers().stream()
-            .filter(
-                d ->
-                    d.getName().isPresent()
-                        || d.getOrganization().isPresent() && d.getEmail().isEmpty())
-            .findFirst()
-            .flatMap(SpdxDocumentBuilder::findPackageSupplierFromDeveloper);
-
-    var packageSupplier =
-        Stream.of(organizationName, developersOrganization, supplierFromDeveloper)
-            .filter(Optional::isPresent)
-            .findFirst()
-            .flatMap(v -> v);
-
-    return packageSupplier.orElse("Organization: NOASSERTION");
-  }
-
-  private static <T> Collector<T, Set<T>, Optional<T>> uniqueOrEmpty() {
-    return Collector.of(
-        java.util.HashSet::new,
-        Set::add,
-        (left, right) -> {
-          left.addAll(right);
-          return left;
-        },
-        set -> set.size() == 1 ? Optional.ofNullable(set.iterator().next()) : Optional.empty());
-  }
-
-  private static Optional<String> findPackageSupplierFromDeveloper(PomInfo.DeveloperInfo d) {
-    Optional<String> name = d.getName();
-    Optional<String> email = d.getEmail();
-    Optional<String> organization = d.getOrganization();
-
-    if (name.isPresent()) {
-      return email
-          .map(s -> String.format("Person: %s (%s)", name.get(), s))
-          .or(() -> Optional.of(String.format("Person: %s", name.get())));
-    } else if (organization.isPresent()) {
-      return Optional.of(String.format("Organization: %s", organization.get()));
-    }
-
     return Optional.empty();
   }
